feat(sprint-79): LLM Infrastructure — AVX2 BitNet kernel, BPE tokenizer, Trinity MoE stub, u32 vocab, QEMU loader pipeline
- 3 new files: bitnet_avx2.rs (AVX2 ternary matmul), trinity.rs (MoE Router stub), bpe.rs (BPE tokenizer HuggingFace JSON parser)
- cortex.rs: vocab_size u16->u32, BPE auto-init in load_model(), dynamic TransformerModel, vectorial RMSNorm
- main.rs: ramdisk loading + QEMU loader fallback at phys 4GB (0x100000000)
- gguf.rs: vocab_size() returns u32
- tensor.rs: matmul_hybrid() delegates to AVX2 kernel
- download_bitnet.py: .bitnet v2 header fixed (u32 vocab, ffn_dim), BPE extraction
- Model: BitNet-b1.58 850M downloaded + converted to .bitnet v2 (1,464 MB)
- Blocker: GQA + BitFFN grouped projections not supported (Sprint 80)

+++ crates/neural-kernel/src/gguf.rs
@@ -413,7 +413,7 @@ impl GgufBackedModel {
             rms_final,
             unembed,
             medusa_heads: Vec::new(),
-            vocab_size: crate::cortex::VOCAB_SIZE,
+            vocab_size: crate::cortex::VOCAB_SIZE as u32,
             hidden: self.hidden_dim,
             num_layers: self.n_layers,
             max_seq: crate::cortex::MAX_SEQ,
@@ -438,8 +438,8 @@ impl Model for GgufBackedModel {
         self.hidden_dim
     }
 
-    fn vocab_size(&self) -> u16 {
-        crate::cortex::VOCAB_SIZE
+    fn vocab_size(&self) -> u32 {
+        crate::cortex::VOCAB_SIZE as u32
     }
 
     fn max_seq(&self) -> usize {
